refactor: move activatedjob interface
* Move the ActivatedJob interface to the protocol-impl module. This causes a refactoring in
  several classes in the broker and engine modules.
* Rename the record() method to jobRecord to be consistent with the jobKey method of the
  same interface.
* The ActivatedJob interface will be later used to stream out jobs to clients, so it is better
  to place it in the protocol-impl module.
