+++ broker/src/main/java/io/camunda/zeebe/broker/jobstream/RemoteJobStream.java
@@ -7,9 +7,9 @@
  */
 package io.camunda.zeebe.broker.jobstream;
 
-import io.camunda.zeebe.engine.processing.streamprocessor.ActivatedJob;
 import io.camunda.zeebe.engine.processing.streamprocessor.JobActivationProperties;
 import io.camunda.zeebe.engine.processing.streamprocessor.JobStreamer.JobStream;
+import io.camunda.zeebe.protocol.impl.stream.job.ActivatedJob;
 import io.camunda.zeebe.transport.stream.api.RemoteStream;
 
 final class RemoteJobStream implements JobStream {
