+++ broker/src/main/java/io/camunda/zeebe/broker/jobstream/RemoteJobStreamer.java
@@ -8,9 +8,9 @@
 package io.camunda.zeebe.broker.jobstream;
 
 import io.atomix.cluster.messaging.ClusterEventService;
-import io.camunda.zeebe.engine.processing.streamprocessor.ActivatedJob;
 import io.camunda.zeebe.engine.processing.streamprocessor.JobActivationProperties;
 import io.camunda.zeebe.engine.processing.streamprocessor.JobStreamer;
+import io.camunda.zeebe.protocol.impl.stream.job.ActivatedJob;
 import io.camunda.zeebe.transport.stream.api.RemoteStreamer;
 import java.util.Optional;
 import org.agrona.DirectBuffer;
